feature: updating tests by refactoring for old player
